Made GUI Universal
the GUI wasn't looking good nor working properly on MacOS. New script uses customtkinter to make the GUI universal
