+++ 3_body_chaotic.csv
@@ -1,0 +1,3 @@
+1.000000000000000000e+00,0.000000000000000000e+00,0.000000000000000000e+00,0.000000000000000000e+00,1.700000000000000122e-02,0.000000000000000000e+00,5.000000000000000000e-01,4.589580000000000000e+05
+0.000000000000000000e+00,0.000000000000000000e+00,0.000000000000000000e+00,0.000000000000000000e+00,1.299999999999999940e-02,0.000000000000000000e+00,1.000000000000000000e+00,1.010981210000000000e+09
+0.000000000000000000e+00,5.000000000000000000e-01,0.000000000000000000e+00,-1.299999999999999940e-02,0.000000000000000000e+00,0.000000000000000000e+00,8.000000000000000444e-01,2.086220500000000000e+07
