+++ solar_system_2002kl6.csv
@@ -1,0 +1,10 @@
+1.464030932887308E-01, -4.242903577346143E-01, -4.790463133878178E-02,  2.100928051306756E-02,  1.046939210833736E-02, -1.072242866521295E-03,1.65956463E-07,6.34E+05
+3.679243047380908E-01, -6.301721435489528E-01, -2.983163202986973E-02,  1.736221586325941E-02,  1.008788583710223E-02, -8.636328856661527E-04,2.44699613E-06,1.05E+07
+4.060102614772616E-02,  9.822623103031199E-01, -1.702056987463614E-04, -1.747399966755353E-02,  6.042032334844135E-04, -5.031326897440181E-07,3.00274E-6,2.07E+07
+1.308018103275883E+00, -4.521223014362638E-01, -4.162598716094207E-02,  5.112722186866305E-03,  1.442784074045314E-02,  1.767100548010790E-04,9.54588E-07,3.08E+07
+-3.660530637172412E+00,  3.851917988321219E+00,  6.583560140608630E-02, -5.560845726100131E-03, -4.843234650469761E-03,  1.445078502690516E-04,0.000954588,4.04E+07
+-5.457736203742084E+00, -8.313348936882264E+00,  3.617578529402181E-01,  4.359579924972715E-03, -3.077424919650049E-03, -1.199546995748008E-04,0.000285716656,5.05E+07
+1.931924877927671E+01,  5.207151660799827E+00, -2.309483093589554E-01, -1.052198361071767E-03,  3.614230679684110E-03,  2.707474657211468E-05,4.36430044E-05,6.06E+07
+2.751503091306775E+01, -1.187857893338104E+01, -3.894960909522387E-01,  1.222909094998878E-03,  2.900646853073834E-03, -8.761030648752844E-05,5.14855965E-05,8.04E+07
+2.792162821653778E-03, -9.178428578803913E-04, -1.359371564679907E-04,  4.050880045372235E-06,  5.163453674383495E-06, -9.812433090995502E-08,1.00E+00,1.09E+09
+-2.568680196003664E+00,  2.454307151069250E+00, -1.964756433689624E-01, -3.759068082351904E-03, -4.893590744502872E-03,  1.106794712579486E-04, 0.0, 16.0
